refactor(cicd): share the GitHub transport between the two release scripts
release-qa-status was forked from gather-release-data and still carried a
verbatim copy of its GitHub transport: onThrottle, createOctokit,
parseRepo, listStandardReleaseTags, findPreviousTag, fetchCommitRange,
resolvePRNumbers and the CommitInfo/ReleaseRef types. ~184 lines,
differing only in comments, brace style and one error message.

Every maintenance change to that layer since the fork has had to patch
both copies — #37140 (for #37138) and #37213 (for #37201), both of them
bug fixes for defects that shipped twice. release-qa-status even carried
a findPreviousTag test whose stated purpose was "guards against drift
from gather-release-data"; the shared module makes that drift
structurally impossible, so the test is gone rather than duplicated.

What moved is plumbing only. fetchPRDetails stays in both callers: the
QA report needs url/author/authorType/externalRefs, the changelog needs
a truncated body and no author, and merging them means a union return
type with each caller ignoring half of it.

Mechanics:
- npm workspaces at .github/scripts — one install, one lockfile instead
  of two, and @octokit/* resolves from the shared module.
- Consumers import ../../shared/github by relative path, so there is no
  build-ordering step between packages.
- gather-release-data now runs through ts-node like release-qa-status
  already did. That drops a build step from the release-notes workflow
  and removes the rootDir constraint that blocks importing ../shared;
  tsc stays as a --noEmit type-check gate.

Verified: gather-release-data output over v26.08.24-01...v26.09.02-01 is
byte-identical to the pre-refactor run, release-qa-status renders its
markdown report, 73 tests pass across the three packages.

Closes: #37382
