+++ .github/scripts/release-qa-status/src/github.ts
@@ -1,208 +1,17 @@
 /**
- * GitHub API interactions: tag resolution, commit range, PR + issue fetching.
+ * PR and issue detail fetching for the QA-status report.
  *
- * Mirrors the patterns in .github/scripts/gather-release-data/src/github.ts so
- * the two tools stay easy to consolidate later.
+ * Auth, throttling, tag listing, range resolution and commit→PR resolution
+ * live in ../../shared/github.ts. This file used to carry a verbatim copy of
+ * all of it, which is how #37138 and #37201 each had to be fixed twice.
+ *
+ * What stays here is what the QA report's shape dictates: PRs carry author
+ * and URL for the Slack digest, and closing refs are split into same-repo vs
+ * external because an external ref cannot be QA-verified in this repo.
  */
 
 import { Octokit } from '@octokit/rest';
-import { throttling } from '@octokit/plugin-throttling';
-import { CommitInfo, ExternalRef, LinkedIssueInfo, PRDetails } from './types';
-
-const STANDARD_RELEASE_PATTERN = /^v\d{2}\.\d{2}\.\d{2}-\d{1,2}$/;
-
-/**
- * resolvePRNumbers is one request per commit — hundreds per release under merge
- * commits. Inside the 5000/hr primary budget, but enough to trip the per-minute
- * secondary limit, where a 403 would degrade into a silently dropped PR.
- */
-const ThrottledOctokit = Octokit.plugin(throttling);
-
-const MAX_RATE_LIMIT_RETRIES = 3;
-
-export function onThrottle(kind: string) {
-  return (
-    retryAfter: number,
-    options: { method?: string; url?: string },
-    _octokit: unknown,
-    retryCount: number
-  ): boolean => {
-    const willRetry = retryCount < MAX_RATE_LIMIT_RETRIES;
-    process.stderr.write(
-      `${kind} rate limit on ${options.method} ${options.url}; ` +
-        (willRetry
-          ? `retry ${retryCount + 1}/${MAX_RATE_LIMIT_RETRIES} in ${retryAfter}s\n`
-          : `giving up after ${MAX_RATE_LIMIT_RETRIES} retries\n`)
-    );
-    return willRetry;
-  };
-}
-
-export function createOctokit(): Octokit {
-  const token = process.env.GITHUB_TOKEN || process.env.GH_TOKEN;
-  if (!token) {
-    throw new Error(
-      'GitHub token required. Set GITHUB_TOKEN or GH_TOKEN (e.g. `export GH_TOKEN=$(gh auth token)`).'
-    );
-  }
-  return new ThrottledOctokit({
-    auth: token,
-    throttle: {
-      onRateLimit: onThrottle('Primary'),
-      onSecondaryRateLimit: onThrottle('Secondary'),
-    },
-  });
-}
-
-export function parseRepo(fullRepo: string): { owner: string; repo: string } {
-  const [owner, repo] = fullRepo.split('/');
-  if (!owner || !repo) {
-    throw new Error(`Invalid repo format: "${fullRepo}". Expected "owner/repo".`);
-  }
-  return { owner, repo };
-}
-
-/** A published standard release, and whether its notes were ever written. */
-export interface ReleaseRef {
-  tag: string;
-  /** False when the release body is empty — a cut whose pipeline died before notes. */
-  hasNotes: boolean;
-}
-
-/**
- * List standard releases (matching vYY.MM.DD-NN pattern), newest first.
- *
- * Kept in sync with gather-release-data/src/github.ts: both resolve the same release
- * boundary, and a divergence puts the QA status and the changelog on different ranges.
- *
- * Drafts are skipped: listReleases returns them first regardless of date, so they
- * corrupt the ordering this function promises, and a draft may duplicate a real tag.
- */
-export async function listStandardReleaseTags(
-  octokit: Octokit,
-  owner: string,
-  repo: string
-): Promise<ReleaseRef[]> {
-  const releases: ReleaseRef[] = [];
-  for await (const response of octokit.paginate.iterator(
-    octokit.repos.listReleases,
-    { owner, repo, per_page: 100 }
-  )) {
-    for (const release of response.data) {
-      if (release.draft) continue;
-      if (STANDARD_RELEASE_PATTERN.test(release.tag_name)) {
-        releases.push({
-          tag: release.tag_name,
-          hasNotes: (release.body ?? '').trim().length > 0,
-        });
-      }
-    }
-  }
-  return releases;
-}
-
-/**
- * Find the previous documented release before `currentTag`.
- *
- * Undocumented releases are skipped: a cut whose pipeline died before writing notes
- * never shipped a QA status either, so stopping at its tag would report on one attempt
- * instead of the whole release.
- */
-export function findPreviousTag(
-  releases: ReleaseRef[],
-  currentTag: string
-): string | undefined {
-  const idx = releases.findIndex((r) => r.tag === currentTag);
-  if (idx === -1) return undefined;
-  return releases.slice(idx + 1).find((r) => r.hasNotes)?.tag;
-}
-
-export async function fetchCommitRange(
-  octokit: Octokit,
-  owner: string,
-  repo: string,
-  fromTag: string,
-  toTag: string
-): Promise<{ totalCommits: number; commits: CommitInfo[] }> {
-  const initial = await octokit.repos.compareCommitsWithBasehead({
-    owner,
-    repo,
-    basehead: `${fromTag}...${toTag}`,
-    per_page: 1,
-  });
-
-  const totalCommits = initial.data.total_commits;
-  if (totalCommits === 0) return { totalCommits: 0, commits: [] };
-
-  const commits: CommitInfo[] = [];
-  const perPage = 250;
-  let page = 1;
-
-  while (commits.length < totalCommits) {
-    const response = await octokit.repos.compareCommitsWithBasehead({
-      owner,
-      repo,
-      basehead: `${fromTag}...${toTag}`,
-      per_page: perPage,
-      page,
-    });
-
-    for (const c of response.data.commits) {
-      commits.push({ sha: c.sha });
-    }
-
-    if (response.data.commits.length < perPage) break;
-    page++;
-  }
-
-  return { totalCommits, commits };
-}
-
-/** Mirrors resolvePRNumbers in gather-release-data/src/github.ts. */
-export async function resolvePRNumbers(
-  octokit: Octokit,
-  owner: string,
-  repo: string,
-  commits: CommitInfo[]
-): Promise<number[]> {
-  const prNumbers = new Set<number>();
-  const BATCH_SIZE = 15;
-
-  for (let i = 0; i < commits.length; i += BATCH_SIZE) {
-    const batch = commits.slice(i, i + BATCH_SIZE);
-
-    const promises = batch.map(async (commit) => {
-      try {
-        const { data } = await octokit.repos.listPullRequestsAssociatedWithCommit({
-          owner,
-          repo,
-          commit_sha: commit.sha,
-          // A default-branch commit resolves to the one merged PR that introduced
-          // it, so page 1 always suffices; this just clears the default 30.
-          per_page: 100,
-        });
-        // Only merged PRs: for commits not reachable from the default branch the
-        // endpoint also returns open PRs.
-        return data.filter((pr) => pr.merged_at).map((pr) => pr.number);
-      } catch (error: unknown) {
-        const errMsg = error instanceof Error ? error.message : String(error);
-        process.stderr.write(
-          `Warning: could not resolve PR for commit ${commit.sha}: ${errMsg}\n`
-        );
-        return [];
-      }
-    });
-
-    const batchResults = await Promise.all(promises);
-    for (const numbers of batchResults) {
-      for (const n of numbers) prNumbers.add(n);
-    }
-
-    if (i + BATCH_SIZE < commits.length) await sleep(500);
-  }
-
-  return Array.from(prNumbers);
-}
+import { ExternalRef, LinkedIssueInfo, PRDetails } from './types';
 
 interface ClosingRefsByPR {
   sameRepo: number[];
